refactor: organize codebase into folders and re-export
