+++ src/transfers/index.ts
@@ -1,0 +1,2 @@
+// Re-export them so we can keep them organized in separate files
+export { recordTransfer } from './transfers'
